// src/interfaces/Character.types.ts
export interface Character {
    sex: string;
    height: string;
    weight: string;
    hairColor: string;
    eyeColor: string;
    personalityTraits: string[];
    skills: string[];
    posture: string;
    healthCondition: string;
    values: string[];
    fears: string[];
    motivations: string[];
    financialStatus: string;
    culture: string;
}